+++ src/test/codemods/__fixtures__/convert-to-let.check.js
@@ -1,6 +1,10 @@
-/*$ { fixture: 'should-trigger', expected: true } $*/
+/*$ { fixture: 'should-trigger-on-var', expected: true } $*/
 
 var a = 'foo'; /*# { pos: 3 } #*/
+
+/*$ { fixture: 'should-trigger-on-const', expected: true } $*/
+
+const a = 'foo'; /*# { pos: 3 } #*/
 
 /*$ { fixture: 'should-not-trigger', expected: false } $*/