+++ src/test/codemods/__fixtures__/convert-to-let.input.js
@@ -1,3 +1,7 @@
-/*$ { fixture: 'should-convert' } $*/
+/*$ { fixture: 'should-convert-var' } $*/
 
 var a = 'foo'; /*# { pos: 3 } #*/
+
+/*$ { fixture: 'should-convert-const' } $*/
+
+const a = 'foo'; /*# { pos: 3 } #*/