+++ src/test/codemods/__fixtures__/convert-to-let.output.js
@@ -1,3 +1,7 @@
-/*$ { fixture: 'should-convert' } $*/
+/*$ { fixture: 'should-convert-var' } $*/
+
+let a = 'foo';
+
+/*$ { fixture: 'should-convert-const' } $*/
 
 let a = 'foo';